XWIKI-14507: Allow actions in the event stream to be observed

+++ xwiki-platform-core/xwiki-platform-activitystream/xwiki-platform-activitystream-api/src/main/java/com/xpn/xwiki/plugin/activitystream/eventstreambridge/BridgeEventStream.java
@@ -78,6 +78,11 @@ public class BridgeEventStream implements EventStream
     public void addEvent(Event e)
     {
         try {
+            // If we are in a loop, don’t add the event
+            if (this.execution.getContext().hasProperty(EVENT_LOOP_CONTEXT_LOCK_PROPERTY)) {
+                return;
+            }
+
             XWikiContext context = getXWikiContext();
             ActivityStreamPlugin plugin = getPlugin(context);
             plugin.getActivityStream().addActivityEvent(eventConverter.convertEventToActivity(e), context);
@@ -91,6 +96,11 @@ public class BridgeEventStream implements EventStream
     public void deleteEvent(Event e)
     {
         try {
+            // Same as addEvent(Event)
+            if (this.execution.getContext().hasProperty(EVENT_LOOP_CONTEXT_LOCK_PROPERTY)) {
+                return;
+            }
+
             XWikiContext context = getXWikiContext();
             ActivityStreamPlugin plugin = getPlugin(context);
             plugin.getActivityStream().deleteActivityEvent(eventConverter.convertEventToActivity(e), context);
@@ -103,12 +113,11 @@ public class BridgeEventStream implements EventStream
     private void sendEventStreamEvent(AbstractEventStreamEvent eventStreamEvent, Event event)
     {
         // In order to avoid infinite loop caused by observation events triggering event stream events that are
-        // themselves triggering new observation events …, we add a new property to the execution context. Therefore,
-        // an observation event declared out of an event stream event can only be triggered once.
-        if (!this.execution.getContext().hasProperty(EVENT_LOOP_CONTEXT_LOCK_PROPERTY)) {
-            this.execution.getContext().newProperty(EVENT_LOOP_CONTEXT_LOCK_PROPERTY).declare();
-            this.observationManager.notify(eventStreamEvent, event);
-        }
+        // themselves triggering new observation events …, the EVENT_LOOP_CONTEXT_LOCK_PROPERTY should be checked
+        // if it is defined in the current execution context before registering an event and sending notifications.
+        // Therefore, an observation event declared out of an event stream event can only be triggered once.
+        this.execution.getContext().newProperty(EVENT_LOOP_CONTEXT_LOCK_PROPERTY).declare();
+        this.observationManager.notify(eventStreamEvent, event);
     }
 
     @Override
